pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "text-to-sql-fine-tuning"
version = "0.1.0"
description = "Fine-tuning LLMs for text-to-SQL using GRPO"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [
    {name = "chrisjcc"},
]
keywords = ["text-to-sql", "fine-tuning", "llm", "grpo", "reinforcement-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "datasets>=2.14.0",
    "accelerate>=0.25.0",
    "trl>=0.7.0",
    "peft>=0.7.0",
    "bitsandbytes>=0.41.0",
    "verifiers>=0.1.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "python-dotenv>=1.0.0",
    "wandb>=0.16.0",
    "sqlparse>=0.4.4",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/chrisjcc/text-to-sql-fine-tuning"
Repository = "https://github.com/chrisjcc/text-to-sql-fine-tuning"
"Bug Tracker" = "https://github.com/chrisjcc/text-to-sql-fine-tuning/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=term-missing"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
